refactor Rank
