+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/calcite/Rank.scala
@@ -19,13 +19,13 @@
 package org.apache.flink.table.plan.nodes.calcite
 
 import org.apache.flink.table.api.TableException
+import org.apache.flink.table.plan.nodes.calcite.RankType.RankType
 import org.apache.flink.table.plan.util._
 
 import org.apache.calcite.plan.{RelOptCluster, RelOptCost, RelOptPlanner, RelTraitSet}
 import org.apache.calcite.rel.`type`.RelDataType
 import org.apache.calcite.rel.metadata.RelMetadataQuery
 import org.apache.calcite.rel.{RelCollation, RelNode, RelWriter, SingleRel}
-import org.apache.calcite.sql.SqlRankFunction
 import org.apache.calcite.sql.`type`.SqlTypeName
 import org.apache.calcite.util.{ImmutableBitSet, NumberUtil}
 
@@ -34,43 +34,41 @@ import java.util
 import scala.collection.JavaConversions._
 
 /**
-  * Relational expression that returns the rows in which the rank function value of each row
+  * Relational expression that returns the rows in which the rank number of each row
   * is in the given range.
   *
-  * <p>NOTES: Different from [[org.apache.calcite.sql.fun.SqlStdOperatorTable.RANK]],
-  * [[Rank]] is a Relational expression， not a window function.
+  * The node is an optimization of `OVER` for some special cases,
+  * e.g.
+  * {{{
+  * SELECT * FROM (
+  *  SELECT a, b, RANK() OVER (PARTITION BY b ORDER BY c) rk FROM MyTable) t
+  * WHERE rk < 10
+  * }}}
+  * can be converted to this node.
   *
-  * <p>[[Rank]] will output rank function value as its last column.
-  *
-  * <p>This RelNode only handles single rank function, is an optimization for some cases. e.g.
-  * <ol>
-  * <li>
-  *   single rank function (on `OVER`) with filter in a SQL query statement
-  * </li>
-  * <li>
-  *   `ORDER BY` with `LIMIT` in a SQL query statement
-  *   (equivalent to `ROW_NUMBER` with filter and project)
-  * </li>
-  * </ol>
-  *
-  * @param cluster        cluster that this relational expression belongs to
-  * @param traitSet       the traits of this rel
-  * @param input          input relational expression
-  * @param rankFunction   rank function, including: CUME_DIST, DENSE_RANK, PERCENT_RANK, RANK,
-  *                       ROW_NUMBER
-  * @param partitionKey   partition keys (may be empty)
-  * @param sortCollation  order keys for rank function
-  * @param rankRange      the expected range of rank function value
+  * @param cluster          cluster that this relational expression belongs to
+  * @param traitSet         the traits of this rel
+  * @param input            input relational expression
+  * @param partitionKey     partition keys (may be empty)
+  * @param orderKey         order keys (should not empty)
+  * @param rankType         rank type to define how exactly generate rank number
+  * @param rankRange        the expected range of rank number value
+  * @param outputRankNumber whether output rank number
   */
 abstract class Rank(
     cluster: RelOptCluster,
     traitSet: RelTraitSet,
     input: RelNode,
-    val rankFunction: SqlRankFunction,
     val partitionKey: ImmutableBitSet,
-    val sortCollation: RelCollation,
-    val rankRange: RankRange)
+    val orderKey: RelCollation,
+    val rankType: RankType,
+    val rankRange: RankRange,
+    val outputRankNumber: Boolean)
   extends SingleRel(cluster, traitSet, input) {
+
+  if (orderKey.getFieldCollations.isEmpty) {
+    throw new TableException("orderKey should not be empty.")
+  }
 
   rankRange match {
     case r: ConstantRankRange =>
@@ -91,10 +89,13 @@ abstract class Rank(
   }
 
   override def deriveRowType(): RelDataType = {
+    if (!outputRankNumber) {
+      return input.getRowType
+    }
     val typeFactory = cluster.getRexBuilder.getTypeFactory
     val typeBuilder = typeFactory.builder()
     input.getRowType.getFieldList.foreach(typeBuilder.add)
-    // rank function column is always the last column, and its type is BIGINT NOT NULL
+    // rank number column is always the last column, and its type is BIGINT NOT NULL
     val allFieldNames = new util.HashSet[String]()
     allFieldNames.addAll(input.getRowType.getFieldNames)
     val rankFieldName = FlinkRelOptUtil.buildUniqueFieldName(allFieldNames, "rk")
@@ -108,10 +109,10 @@ abstract class Rank(
       case (name, idx) => s"$name=$$$idx"
     }.mkString(", ")
     super.explainTerms(pw)
-      .item("rankFunction", rankFunction)
+      .item("rankType", rankType)
       .item("rankRange", rankRange.toString())
       .item("partitionBy", partitionKey.map(i => s"$$$i").mkString(","))
-      .item("orderBy", RelExplainUtil.collationToString(sortCollation))
+      .item("orderBy", RelExplainUtil.collationToString(orderKey))
       .item("select", select)
   }
 
@@ -137,6 +138,37 @@ abstract class Rank(
     planner.getCostFactory.makeCost(rowCount, cpuCost, 0)
   }
 
+}
+
+/**
+  * An enumeration of rank type, usable to tell the [[Rank]] node how exactly generate rank number.
+  */
+object RankType extends Enumeration {
+  type RankType = Value
+
+  /**
+    * Returns a unique sequential number for each row within the partition based on the order,
+    * starting at 1 for the first row in each partition and without repeating or skipping
+    * numbers in the ranking result of each partition. If there are duplicate values within the
+    * row set, the ranking numbers will be assigned arbitrarily.
+    */
+  val ROW_NUMBER: RankType.Value = Value
+
+  /**
+    * Returns a unique rank number for each distinct row within the partition based on the order,
+    * starting at 1 for the first row in each partition, with the same rank for duplicate values
+    * and leaving gaps between the ranks; this gap appears in the sequence after the duplicate
+    * values.
+    */
+  val RANK: RankType.Value = Value
+
+  /**
+    * is similar to the RANK by generating a unique rank number for each distinct row
+    * within the partition based on the order, starting at 1 for the first row in each partition,
+    * ranking the rows with equal values with the same rank number, except that it does not skip
+    * any rank, leaving no gaps between the ranks.
+    */
+  val DENSE_RANK: RankType.Value = Value
 }
 
 sealed trait RankRange extends Serializable {
