+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/physical/stream/StreamExecLimit.scala
@@ -15,11 +15,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.apache.flink.table.plan.nodes.physical.stream
 
-import org.apache.flink.annotation.Experimental
-import org.apache.flink.table.plan.util.RelExplainUtil
+import org.apache.flink.table.plan.util.{FlinkRelOptUtil, RelExplainUtil}
 
 import org.apache.calcite.plan.{RelOptCluster, RelTraitSet}
 import org.apache.calcite.rel._
@@ -29,28 +27,26 @@ import org.apache.calcite.rex.RexNode
 /**
   * Stream physical RelNode for [[Sort]].
   *
-  * <b>NOTES:</b> This class is used for testing with bounded source now.
-  * If a query is converted to this node in product environment, an exception will be thrown.
-  *
-  * @see [[StreamExecTemporalSort]] which must be time-ascending-order sort without `limit`.
-  *
-  * e.g.
-  * ''SELECT * FROM TABLE ORDER BY ROWTIME, a'' will be converted to [[StreamExecTemporalSort]]
-  * ''SELECT * FROM TABLE ORDER BY a, ROWTIME'' will be converted to [[StreamExecSort]]
+  * This node will output `limit` records beginning with the first `offset` records without sort.
   */
-@Experimental
-class StreamExecSort(
+class StreamExecLimit(
     cluster: RelOptCluster,
     traitSet: RelTraitSet,
     inputRel: RelNode,
-    sortCollation: RelCollation)
-  extends Sort(cluster, traitSet, inputRel, sortCollation)
+    offset: RexNode,
+    fetch: RexNode)
+  extends Sort(
+    cluster,
+    traitSet,
+    inputRel,
+    RelCollations.EMPTY,
+    offset,
+    fetch)
   with StreamPhysicalRel {
 
-  /**
-    * this node will not produce or consume retraction message
-    * due to it starts sending data to output after all input data has come.
-    */
+  private lazy val limitStart: Long = FlinkRelOptUtil.getLimitStart(offset)
+  private lazy val limitEnd: Long = FlinkRelOptUtil.getLimitEnd(offset, fetch)
+
   override def producesUpdates: Boolean = false
 
   override def needsUpdatesAsRetraction(input: RelNode): Boolean = false
@@ -63,16 +59,17 @@ class StreamExecSort(
 
   override def copy(
       traitSet: RelTraitSet,
-      input: RelNode,
+      newInput: RelNode,
       newCollation: RelCollation,
       offset: RexNode,
       fetch: RexNode): Sort = {
-    new StreamExecSort(cluster, traitSet, input, newCollation)
+    new StreamExecLimit(cluster, traitSet, newInput, offset, fetch)
   }
 
   override def explainTerms(pw: RelWriter): RelWriter = {
-    pw.input("input", getInput())
-      .item("orderBy", RelExplainUtil.collationToString(sortCollation, getRowType))
+    pw.input("input", getInput)
+      .item("offset", limitStart)
+      .item("fetch", RelExplainUtil.fetchToString(fetch))
   }
 
 }
