+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/rules/FlinkStreamRuleSets.scala
@@ -227,8 +227,9 @@ object FlinkStreamRuleSets {
     StreamExecCalcRule.INSTANCE,
     StreamExecUnionRule.INSTANCE,
     StreamExecSortRule.INSTANCE,
-    StreamExecRankRule.SORT_INSTANCE,
-    StreamExecRankRule.RANK_INSTANCE,
+    StreamExecLimitRule.INSTANCE,
+    StreamExecSortLimitRule.INSTANCE,
+    StreamExecRankRule.INSTANCE,
     StreamExecTemporalSortRule.INSTANCE,
     StreamExecFirstLastRowRule.SORT_INSTANCE,
     StreamExecFirstLastRowRule.RANK_INSTANCE,
