+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/rules/physical/batch/BatchExecLimitRule.scala
@@ -41,7 +41,7 @@ import org.apache.calcite.sql.`type`.SqlTypeName
   * }}}
   * when fetch is not null, or
   * {{{
-  * BatchExecLimit (global)
+  * BatchExecLimit
   * +- BatchExecExchange (singleton)
   *    +- input of sort
   * }}}
