+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/rules/physical/stream/StreamExecFirstLastRowRule.scala
@@ -21,9 +21,9 @@ package org.apache.flink.table.plan.rules.physical.stream
 import org.apache.flink.table.calcite.FlinkTypeFactory
 import org.apache.flink.table.plan.`trait`.FlinkRelDistribution
 import org.apache.flink.table.plan.nodes.FlinkConventions
-import org.apache.flink.table.plan.nodes.calcite.ConstantRankRange
+import org.apache.flink.table.plan.nodes.calcite.{ConstantRankRange, RankType}
 import org.apache.flink.table.plan.nodes.logical.{FlinkLogicalRank, FlinkLogicalSort}
-import org.apache.flink.table.plan.nodes.physical.stream.{StreamExecFirstLastRow, StreamExecRank}
+import org.apache.flink.table.plan.nodes.physical.stream.{StreamExecFirstLastRow, StreamExecRank, StreamExecSortLimit}
 import org.apache.flink.table.plan.util.FlinkRelOptUtil
 
 import org.apache.calcite.plan.{RelOptRule, RelOptRuleCall}
@@ -31,15 +31,14 @@ import org.apache.calcite.rel.`type`.RelDataType
 import org.apache.calcite.rel.convert.ConverterRule
 import org.apache.calcite.rel.core.Sort
 import org.apache.calcite.rel.{RelCollation, RelNode}
-import org.apache.calcite.sql.SqlKind
 
 /**
   * Rule that matches [[FlinkLogicalSort]] which is sorted by proc-time attribute and
   * fetches only one record started from 0, and converts it to [[StreamExecFirstLastRow]].
   *
   * NOTES: Queries that can be converted to [[StreamExecFirstLastRow]] could be converted to
-  * [[StreamExecRank]] too. [[StreamExecFirstLastRow]] is more efficient than [[StreamExecRank]]
-  * due to mini-batch and less state access.
+  * [[StreamExecSortLimit]] too. [[StreamExecFirstLastRow]] is more efficient than
+  * [[StreamExecSortLimit]] due to mini-batch and less state access.
   *
   * e.g.
   * 1. ''SELECT a FROM MyTable ORDER BY proctime LIMIT 1'' will be converted to FirstRow
@@ -85,7 +84,7 @@ class StreamExecFirstLastRowFromSortRule
 
 /**
   * Rule that matches [[FlinkLogicalRank]] which is sorted by proc-time attribute and
-  * limits 1 and its RankFunction is ROW_NUMBER, and converts it to [[StreamExecFirstLastRow]].
+  * limits 1 and its rank type is ROW_NUMBER, and converts it to [[StreamExecFirstLastRow]].
   *
   * NOTES: Queries that can be converted to [[StreamExecFirstLastRow]] could be converted to
   * [[StreamExecRank]] too. [[StreamExecFirstLastRow]] is more efficient than [[StreamExecRank]]
@@ -121,7 +120,7 @@ class StreamExecFirstLastRowFromRankRule
 
   override def convert(rel: RelNode): RelNode = {
     val rank = rel.asInstanceOf[FlinkLogicalRank]
-    val fieldCollation = rank.sortCollation.getFieldCollations.get(0)
+    val fieldCollation = rank.orderKey.getFieldCollations.get(0)
     val fieldType = rank.getInput.getRowType.getFieldList.get(fieldCollation.getFieldIndex).getType
     val isRowtime = FlinkTypeFactory.isRowtimeIndicatorType(fieldType)
 
@@ -159,10 +158,10 @@ object StreamExecFirstLastRowRule {
     * @return True if the input rank could be converted to [[StreamExecFirstLastRow]]
     */
   def canConvertToFirstLastRow(rank: FlinkLogicalRank): Boolean = {
-    val sortCollation = rank.sortCollation
+    val sortCollation = rank.orderKey
     val rankRange = rank.rankRange
 
-    val isRowNumberFunction = rank.rankFunction.getKind == SqlKind.ROW_NUMBER
+    val isRowNumberType = rank.rankType == RankType.ROW_NUMBER
 
     val isLimit1 = rankRange match {
       case ConstantRankRange(rankStart, rankEnd) => rankStart == 1 && rankEnd == 1
@@ -172,7 +171,7 @@ object StreamExecFirstLastRowRule {
     val inputRowType = rank.getInput.getRowType
     val isSortOnProctime = sortOnProcTimeAttribute(sortCollation, inputRowType)
 
-    !rank.outputRankFunColumn && isLimit1 && isSortOnProctime && isRowNumberFunction
+    !rank.outputRankNumber && isLimit1 && isSortOnProctime && isRowNumberType
   }
 
   /**
