+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/rules/physical/stream/StreamExecRankRule.scala
@@ -18,82 +18,25 @@
 
 package org.apache.flink.table.plan.rules.physical.stream
 
-import org.apache.flink.table.api.TableException
 import org.apache.flink.table.plan.`trait`.FlinkRelDistribution
 import org.apache.flink.table.plan.nodes.FlinkConventions
-import org.apache.flink.table.plan.nodes.calcite.ConstantRankRange
-import org.apache.flink.table.plan.nodes.logical.{FlinkLogicalRank, FlinkLogicalSort}
+import org.apache.flink.table.plan.nodes.logical.FlinkLogicalRank
 import org.apache.flink.table.plan.nodes.physical.stream.{StreamExecFirstLastRow, StreamExecRank}
-import org.apache.flink.table.plan.util.FlinkRelOptUtil
 
 import org.apache.calcite.plan.{RelOptRule, RelOptRuleCall}
 import org.apache.calcite.rel.RelNode
 import org.apache.calcite.rel.convert.ConverterRule
-import org.apache.calcite.rex.RexLiteral
-import org.apache.calcite.sql.fun.SqlStdOperatorTable
-import org.apache.calcite.util.ImmutableBitSet
-
-/**
-  * Rule that converts [[FlinkLogicalSort]] with non-null `fetch` to [[StreamExecRank]].
-  * NOTES: the sort can not be converted to [[StreamExecFirstLastRow]].
-  */
-class StreamExecRankFromSortRule
-  extends ConverterRule(
-    classOf[FlinkLogicalSort],
-    FlinkConventions.LOGICAL,
-    FlinkConventions.STREAM_PHYSICAL,
-    "StreamExecRankFromSortRule") {
-
-  override def matches(call: RelOptRuleCall): Boolean = {
-    val sort: FlinkLogicalSort = call.rel(0)
-    StreamExecRankRule.canConvertToRank(sort) &&
-      !StreamExecFirstLastRowRule.canConvertToFirstLastRow(sort)
-  }
-
-  override def convert(rel: RelNode): RelNode = {
-    val sort = rel.asInstanceOf[FlinkLogicalSort]
-    val input = sort.getInput
-    val requiredDistribution = FlinkRelDistribution.SINGLETON
-
-    val requiredTraitSet = input.getTraitSet
-      .replace(FlinkConventions.STREAM_PHYSICAL)
-      .replace(requiredDistribution)
-
-    val newInput: RelNode = RelOptRule.convert(input, requiredTraitSet)
-
-    val rankStart = FlinkRelOptUtil.getLimitStart(sort.offset) + 1 // rank start always start from 1
-
-    val rankEnd = if (sort.fetch != null) {
-      FlinkRelOptUtil.getLimitEnd(sort.offset, sort.fetch)
-    } else {
-      // we have checked in matches method that fetch is not null
-      throw new TableException("This should never happen, please file an issue.")
-    }
-
-    val providedTraitSet = rel.getTraitSet.replace(FlinkConventions.STREAM_PHYSICAL)
-    new StreamExecRank(
-      rel.getCluster,
-      providedTraitSet,
-      newInput,
-      SqlStdOperatorTable.ROW_NUMBER,
-      ImmutableBitSet.of(),
-      sort.collation,
-      ConstantRankRange(rankStart, rankEnd),
-      outputRankFunColumn = false)
-  }
-}
 
 /**
   * Rule that converts [[FlinkLogicalRank]] with fetch to [[StreamExecRank]].
   * NOTES: the rank can not be converted to [[StreamExecFirstLastRow]].
   */
-class StreamExecRankFromRankRule
+class StreamExecRankRule
   extends ConverterRule(
     classOf[FlinkLogicalRank],
     FlinkConventions.LOGICAL,
     FlinkConventions.STREAM_PHYSICAL,
-    "StreamExecRankFromRankRule") {
-
+    "StreamExecRankRule") {
 
   override def matches(call: RelOptRuleCall): Boolean = {
     val rank: FlinkLogicalRank = call.rel(0)
@@ -104,7 +47,7 @@ class StreamExecRankFromRankRule
     val rank = rel.asInstanceOf[FlinkLogicalRank]
     val input = rank.getInput
     val requiredDistribution = if (!rank.partitionKey.isEmpty) {
-      FlinkRelDistribution.hash(rank.partitionKey.asList())
+      FlinkRelDistribution.hash(rank.partitionKey.toList)
     } else {
       FlinkRelDistribution.SINGLETON
     }
@@ -118,27 +61,14 @@ class StreamExecRankFromRankRule
       rank.getCluster,
       providedTraitSet,
       newInput,
-      rank.rankFunction,
       rank.partitionKey,
-      rank.sortCollation,
+      rank.orderKey,
+      rank.rankType,
       rank.rankRange,
-      rank.outputRankFunColumn)
+      rank.outputRankNumber)
   }
 }
 
 object StreamExecRankRule {
-  val SORT_INSTANCE: RelOptRule = new StreamExecRankFromSortRule
-  val RANK_INSTANCE: RelOptRule = new StreamExecRankFromRankRule
-
-  /**
-    * Whether the given sort could be converted to [[StreamExecRank]].
-    *
-    * Return true if `fetch` is not null and is greater than 0, else false.
-    *
-    * @param sort the [[FlinkLogicalSort]] node
-    * @return True if the input sort could be converted to [[StreamExecRank]]
-    */
-  def canConvertToRank(sort: FlinkLogicalSort): Boolean = {
-    sort.fetch != null && RexLiteral.intValue(sort.fetch) > 0
-  }
+  val INSTANCE: RelOptRule = new StreamExecRankRule
 }
