+++ flink-table/flink-table-planner-blink/src/test/scala/org/apache/flink/table/plan/stream/sql/SortTest.scala
@@ -27,36 +27,17 @@ class SortTest extends TableTestBase {
 
   private val util = streamTestUtil()
   util.addDataStream[(Int, String, Long)]("MyTable", 'a, 'b, 'c, 'proctime, 'rowtime)
-  // TODO optimize `limit 0`
 
   @Test
   def testSortProcessingTime(): Unit = {
+    // be converted to TemporalSort
     util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime, c")
   }
 
   @Test
   def testSortRowTime(): Unit = {
+    // be converted to TemporalSort
     util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime, c")
-  }
-
-  @Test
-  def testSortProcessingTimeWithLimit(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime, c LIMIT 2")
-  }
-
-  @Test
-  def testSortProcessingTimeWithOffset(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime, c OFFSET 2")
-  }
-
-  @Test
-  def testSortRowTimeWithLimit(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime, c LIMIT 2")
-  }
-
-  @Test
-  def testSortRowTimeWithOffset(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime, c OFFSET 2")
   }
 
   @Test
@@ -70,26 +51,6 @@ class SortTest extends TableTestBase {
   }
 
   @Test
-  def testSortProcessingTimeDescWithLimit(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime desc, c LIMIT 2")
-  }
-
-  @Test
-  def testSortProcessingTimeDescWithOffset(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime desc, c OFFSET 2")
-  }
-
-  @Test
-  def testSortRowTimeDescWithLimit(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime desc, c LIMIT 2")
-  }
-
-  @Test
-  def testSortRowTimeDescWithOffset(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime desc, c OFFSET 2")
-  }
-
-  @Test
   def testSortProcessingTimeSecond(): Unit = {
     util.verifyPlan("SELECT a FROM MyTable ORDER BY c, proctime")
   }
@@ -97,26 +58,6 @@ class SortTest extends TableTestBase {
   @Test
   def testSortRowTimeSecond(): Unit = {
     util.verifyPlan("SELECT a FROM MyTable ORDER BY c, rowtime")
-  }
-
-  @Test
-  def testSortProcessingTimeSecondWithLimit(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, proctime LIMIT 2")
-  }
-
-  @Test
-  def testSortProcessingTimeSecondWithOffset(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, proctime OFFSET 2")
-  }
-
-  @Test
-  def testSortRowTimeSecondWithLimit(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, rowtime LIMIT 2")
-  }
-
-  @Test
-  def testSortRowTimeSecondWithOffset(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, rowtime OFFSET 2")
   }
 
   @Test
@@ -130,357 +71,7 @@ class SortTest extends TableTestBase {
   }
 
   @Test
-  def testSortProcessingTimeSecondDescWithLimit(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, proctime desc LIMIT 2")
-  }
-
-  @Test
-  def testSortProcessingTimeSecondDescWithOffset(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, proctime desc OFFSET 2")
-  }
-
-  @Test
-  def testSortRowTimeDescSecondWithLimit(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, rowtime desc LIMIT 2")
-  }
-
-  @Test
-  def testSortRowTimeDescSecondWithOffset(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, rowtime desc OFFSET 2")
-  }
-
-  @Test
-  def testSortProcessingTimeSecondDescWithLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, proctime desc LIMIT 0")
-  }
-
-  @Test
-  def testSortProcessingTimeSecondDescWithLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, proctime desc LIMIT 1")
-  }
-
-  @Test
-  def testSortProcessingTimeSecondAscWithLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, proctime LIMIT 1")
-  }
-
-  @Test
-  def testSortProcessingTimeFirstDescWithLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime desc LIMIT 0")
-  }
-
-  @Test
-  def testSortProcessingTimeFirstDescWithLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime desc LIMIT 1")
-  }
-
-  @Test
-  def testSortProcessingTimeFirstAscWithLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime, c LIMIT 0")
-  }
-
-  @Test
-  def testSortProcessingTimeFirstAscWithLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime, c LIMIT 1")
-  }
-
-  @Test
-  def testSortProcessingTimeFirstDescWithOffSet0AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime, c desc LIMIT 1 OFFSET 0")
-  }
-
-  @Test
-  def testSortProcessingTimeFirstDescWithOffSet1AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime, c desc LIMIT 1 OFFSET 1")
-  }
-
-  @Test
-  def testSortProcessingTimeFirstAscWithOffSet1AndLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime, c LIMIT 0 OFFSET 1")
-  }
-
-  @Test
-  def testSortProcessingTimeFirstAscWithOffSet1AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime, c LIMIT 1 OFFSET 1")
-  }
-
-  @Test
-  def testSortProcessingTimeAscWithLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime LIMIT 1")
-  }
-
-  @Test
-  def testSortProcessingTimeDescWithOffSet1AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime desc LIMIT 1 OFFSET 1")
-  }
-
-  @Test
-  def testSortProcessingTimeAscWithOffSet1AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime LIMIT 1 OFFSET 1")
-  }
-
-  @Test
-  def testSortProcessingTimeAscWithOffSet0AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime LIMIT 1 OFFSET 0")
-  }
-
-  @Test
-  def testSortProcessingTimeDescWithOffSet0AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY proctime desc LIMIT 1 OFFSET 0")
-  }
-
-  @Test
-  def testSortRowTimeSecondDescWithLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, rowtime desc LIMIT 0")
-  }
-
-  @Test
-  def testSortRowTimeSecondDescWithLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, rowtime desc LIMIT 1")
-  }
-
-  @Test
-  def testSortRowTimeSecondAscWithLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, rowtime LIMIT 0")
-  }
-
-  @Test
-  def testSortRowTimeSecondAscWithLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c, rowtime LIMIT 1")
-  }
-
-  @Test
-  def testSortRowTimeFirstDescWithLimit0(): Unit = {
-    // FirstLastRow does not support sort on rowtime now, so it is translated to Rank currently
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime desc LIMIT 0")
-  }
-
-  @Test
-  def testSortRowTimeFirstDescWithLimit1(): Unit = {
-    // FirstLastRow does not support sort on rowtime now, so it is translated to Rank currently
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime desc LIMIT 1")
-  }
-
-  @Test
-  def testSortRowTimeFirstAscWithLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime, c LIMIT 0")
-  }
-
-  @Test
-  def testSortRowTimeFirstAscWithLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime, c LIMIT 1")
-  }
-
-  @Test
-  def testSortRowTimeFirstDescWithOffSet1AndLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime, c desc LIMIT 0 OFFSET 1")
-  }
-
-  @Test
-  def testSortRowTimeFirstDescWithOffSet1AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime, c desc LIMIT 1 OFFSET 1")
-  }
-
-  @Test
-  def testSortRowTimeFirstAscWithOffSet1AndLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime, c LIMIT 0 OFFSET 1")
-  }
-
-  @Test
-  def testSortRowTimeFirstAscWithOffSet1AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime, c LIMIT 1 OFFSET 1")
-  }
-
-  @Test
-  def testSortRowTimeAscWithLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime LIMIT 0")
-  }
-
-  @Test
-  def testSortRowTimeAscWithLimit1(): Unit = {
-    // FirstLastRow does not support sort on rowtime now, so it is translated to Rank currently
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime LIMIT 1")
-  }
-
-  @Test
-  def testSortRowTimeDescWithOffSet1AndLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime desc LIMIT 0 OFFSET 1")
-  }
-
-  @Test
-  def testSortRowTimeDescWithOffSet1AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime desc LIMIT 1 OFFSET 1")
-  }
-
-  @Test
-  def testSortRowTimeAscWithOffSet1AndLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime LIMIT 0 OFFSET 1")
-  }
-
-  @Test
-  def testSortRowTimeAscWithOffSet1AndLimit1(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime LIMIT 1 OFFSET 1")
-  }
-
-  @Test
-  def testSortRowTimeDescWithOffSet0AndLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime desc LIMIT 0 OFFSET 0")
-  }
-
-  @Test
-  def testSortRowTimeDescWithOffSet0AndLimit1(): Unit = {
-    // FirstLastRow does not support sort on rowtime now, so it is translated to Rank currently
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime desc LIMIT 1 OFFSET 0")
-  }
-
-  @Test
-  def testSortRowTimeAscWithOffSet0AndLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime LIMIT 0 OFFSET 0")
-  }
-
-  @Test
-  def testSortRowTimeAscWithOffSet0AndLimit1(): Unit = {
-    // FirstLastRow does not support sort on rowtime now, so it is translated to Rank currently
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY rowtime LIMIT 1 OFFSET 0")
-  }
-
-  @Test
   def testSortWithoutTime(): Unit = {
     util.verifyPlan("SELECT a FROM MyTable ORDER BY c")
-  }
-
-  @Test
-  def testSortWithoutTimeWithLimit(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c LIMIT 2")
-  }
-
-  @Test
-  def testSortWithoutTimeWithLimit0(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c LIMIT 0")
-  }
-
-  @Test
-  def testSortWithoutTimeWithOffset(): Unit = {
-    util.verifyPlan("SELECT a FROM MyTable ORDER BY c OFFSET 2")
-  }
-
-  @Test
-  def testLimitWithRowTime(): Unit = {
-    util.verifyPlan("SELECT rowtime, c FROM MyTable LIMIT 2")
-  }
-
-  @Test
-  def testLimit0WithRowTime(): Unit = {
-    util.verifyPlan("SELECT rowtime, c FROM MyTable LIMIT 0")
-  }
-
-  @Test
-  def testOffsetWithRowTime(): Unit = {
-    util.verifyPlan("SELECT rowtime, c FROM MyTable OFFSET 2")
-  }
-
-  @Test
-  def testLimit0WithProcessingTime(): Unit = {
-    util.verifyPlan("SELECT proctime, c FROM MyTable LIMIT 0")
-  }
-
-  @Test
-  def testLimitWithProcessingTime(): Unit = {
-    util.verifyPlan("SELECT proctime, c FROM MyTable LIMIT 2")
-  }
-
-  @Test
-  def testOffsetWithProcessingTime(): Unit = {
-    util.verifyPlan("SELECT proctime, c FROM MyTable OFFSET 2")
-  }
-
-  @Test
-  def testLimit0WithRowTimeSecond(): Unit = {
-    util.verifyPlan("SELECT c, rowtime FROM MyTable LIMIT 0")
-  }
-
-  @Test
-  def testLimitWithRowTimeSecond(): Unit = {
-    util.verifyPlan("SELECT c, rowtime FROM MyTable LIMIT 2")
-  }
-
-  @Test
-  def testOffsetWithRowTimeSecond(): Unit = {
-    util.verifyPlan("SELECT c, rowtime FROM MyTable OFFSET 2")
-  }
-
-  @Test
-  def testLimit0WithProcessingTimeSecond(): Unit = {
-    util.verifyPlan("SELECT c, proctime FROM MyTable LIMIT 0")
-  }
-
-  @Test
-  def testLimitWithProcessingTimeSecond(): Unit = {
-    util.verifyPlan("SELECT c, proctime FROM MyTable LIMIT 2")
-  }
-
-  @Test
-  def testOffsetWithProcessingTimeSecond(): Unit = {
-    util.verifyPlan("SELECT c, proctime FROM MyTable OFFSET 2")
-  }
-
-  @Test
-  def testLimit0WithRowTimeDesc(): Unit = {
-    util.verifyPlan("SELECT rowtime desc, c FROM MyTable LIMIT 0")
-  }
-
-  @Test
-  def testLimitWithRowTimeDesc(): Unit = {
-    util.verifyPlan("SELECT rowtime desc, c FROM MyTable LIMIT 2")
-  }
-
-  @Test
-  def testOffsetWithRowTimeDesc(): Unit = {
-    util.verifyPlan("SELECT rowtime desc, c FROM MyTable OFFSET 2")
-  }
-
-  @Test
-  def testLimit0WithProcessingTimeDesc(): Unit = {
-    util.verifyPlan("SELECT proctime desc, c FROM MyTable LIMIT 0")
-  }
-
-  @Test
-  def testLimitWithProcessingTimeDesc(): Unit = {
-    util.verifyPlan("SELECT proctime desc, c FROM MyTable LIMIT 2")
-  }
-
-  @Test
-  def testOffsetWithProcessingTimeDesc(): Unit = {
-    util.verifyPlan("SELECT proctime desc, c FROM MyTable OFFSET 2")
-  }
-
-  @Test
-  def testLimit0WithRowTimeDescSecond(): Unit = {
-    util.verifyPlan("SELECT c, rowtime desc FROM MyTable LIMIT 0")
-  }
-
-  @Test
-  def testLimitWithRowTimeDescSecond(): Unit = {
-    util.verifyPlan("SELECT c, rowtime desc FROM MyTable LIMIT 2")
-  }
-
-  @Test
-  def testOffsetWithRowTimeDescSecond(): Unit = {
-    util.verifyPlan("SELECT c, rowtime desc FROM MyTable OFFSET 2")
-  }
-
-  @Test
-  def testLimit0WithProcessingTimeDescSecond(): Unit = {
-    util.verifyPlan("SELECT c, proctime desc FROM MyTable LIMIT 0")
-  }
-
-  @Test
-  def testLimitWithProcessingTimeDescSecond(): Unit = {
-    util.verifyPlan("SELECT c, proctime desc FROM MyTable LIMIT 2")
-  }
-
-  @Test
-  def testOffsetWithProcessingTimeDescSecond(): Unit = {
-    util.verifyPlan("SELECT c, proctime desc FROM MyTable OFFSET 2")
   }
 }
